move(department): department entity and enum class move to entity package
