+++ src/main/java/com/dongsoop/dongsoop/department/repository/DepartmentRepository.java
@@ -1,7 +1,7 @@
 package com.dongsoop.dongsoop.department.repository;
 
-import com.dongsoop.dongsoop.department.Department;
-import com.dongsoop.dongsoop.department.DepartmentType;
+import com.dongsoop.dongsoop.department.entity.Department;
+import com.dongsoop.dongsoop.department.entity.DepartmentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DepartmentRepository extends JpaRepository<Department, DepartmentType> {
