+++ src/main/java/com/dongsoop/dongsoop/notice/entity/Notice.java
@@ -1,6 +1,6 @@
 package com.dongsoop.dongsoop.notice.entity;
 
-import com.dongsoop.dongsoop.department.Department;
+import com.dongsoop.dongsoop.department.entity.Department;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
