+++ src/main/java/com/dongsoop/dongsoop/notice/service/NoticeService.java
@@ -1,6 +1,6 @@
 package com.dongsoop.dongsoop.notice.service;
 
-import com.dongsoop.dongsoop.department.DepartmentType;
+import com.dongsoop.dongsoop.department.entity.DepartmentType;
 import com.dongsoop.dongsoop.notice.dto.NoticeListResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
