+++ src/main/java/com/dongsoop/dongsoop/notice/service/NoticeServiceImpl.java
@@ -1,7 +1,7 @@
 package com.dongsoop.dongsoop.notice.service;
 
-import com.dongsoop.dongsoop.department.Department;
-import com.dongsoop.dongsoop.department.DepartmentType;
+import com.dongsoop.dongsoop.department.entity.Department;
+import com.dongsoop.dongsoop.department.entity.DepartmentType;
 import com.dongsoop.dongsoop.department.repository.DepartmentRepository;
 import com.dongsoop.dongsoop.exception.domain.department.DepartmentNotFoundException;
 import com.dongsoop.dongsoop.notice.dto.NoticeListResponse;
